+++ CHANGELOG.md
@@ -8,8 +8,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Enable lz4 compression type for Kafka producer
 
+## [2.0.1] - 2017-08-11
+
+### Fixed
+- Added validation of offsets availability when resetting subscription cursors.
+
+### Changed
+- Removed authorization for subscription creation
+
+## [2.0.0] - 2017-08-09
+
 ### Changed
 - Changed imports format to have the same structure
+
+### Removed
+- Removed read_scopes and write_scopes from event types
+- Removed CHECK_APPLICATION_LEVEL_PERMISSIONS feature
 
 ## [1.1.3] - 2017-08-03